feat(base): mutable sorted set implementation.
