+++ packages/base/src/commonMain/kotlin/elide/struct/SortedMaps.kt
@@ -17,6 +17,7 @@ import kotlin.collections.Map.Entry
 import elide.struct.api.MutableSortedMap
 import elide.struct.api.SortedMap
 
+/** A stubbed [SortedMap] implementation, meant to be used as an optimized return value for [emptySortedMap]. */
 private object EmptySortedMap : SortedMap<Nothing, Nothing> {
   override fun get(key: Nothing): Nothing? = null
 
@@ -30,28 +31,43 @@ private object EmptySortedMap : SortedMap<Nothing, Nothing> {
   override fun containsKey(key: Nothing): Boolean = false
 }
 
-// TODO(@darvld): test this
+/** Returns an empty, read-only sorted map. */
 @Suppress("UNCHECKED_CAST")
 public fun <K : Comparable<K>, V> emptySortedMap(): SortedMap<K, V> {
   return EmptySortedMap as SortedMap<K, V>
 }
 
 /**
+ * Returns a new read-only sorted map with the specified contents, given as a list of pairs where the first value is
+ * the key and the second is the value.
  *
+ * If multiple pairs have the same key, the resulting map will contain the value from the last of those pairs.
+ *
+ * The order of the entries is not preserved, but iterating over the map will always yield entries sorted by their keys.
  */
 public fun <K : Comparable<K>, V> sortedMapOf(pairs: Collection<Pair<K, V>>): SortedMap<K, V> {
   return RedBlackTreeMap<K, V>().apply { pairs.forEach { put(it.first, it.second) } }
 }
 
 /**
+ * Returns a new read-only sorted map with the specified contents, given as a list of pairs where the first value is
+ * the key and the second is the value.
  *
+ * If multiple pairs have the same key, the resulting map will contain the value from the last of those pairs.
+ *
+ * The order of the entries is not preserved, but iterating over the map will always yield entries sorted by their keys.
  */
 public fun <K : Comparable<K>, V> sortedMapOf(vararg pairs: Pair<K, V>): SortedMap<K, V> {
   return RedBlackTreeMap<K, V>().apply { pairs.forEach { put(it.first, it.second) } }
 }
 
 /**
+ * Returns a new mutable sorted map with the specified contents, given as a list of pairs where the first value is the
+ * key and the second is the value.
  *
+ * If multiple pairs have the same key, the resulting map will contain the value from the last of those pairs.
+ *
+ * The order of the entries is not preserved, but iterating over the map will always yield entries sorted by their keys.
  */
 public fun <K : Comparable<K>, V> mutableSortedMapOf(vararg pairs: Pair<K, V>): MutableSortedMap<K, V> {
   return RedBlackTreeMap<K, V>().apply { pairs.forEach { put(it.first, it.second) } }
